refactor: const ms package
