+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/const/constimmutable/runConstContainsTests.kt
@@ -1,6 +1,6 @@
-package xyz.lbres.kotlinutils.set.multiset.constimmutable
+package xyz.lbres.kotlinutils.set.multiset.const.constimmutable
 
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.const.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
