+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/const/constmutable/runMutableBinaryTests.kt
@@ -1,6 +1,9 @@
-package xyz.lbres.kotlinutils.set.multiset.constmutable
+package xyz.lbres.kotlinutils.set.multiset.const.constmutable
 
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.const.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.MutableMultiSet
+import xyz.lbres.kotlinutils.set.multiset.emptyMultiSet
+import xyz.lbres.kotlinutils.set.multiset.multiSetOf
 import xyz.lbres.kotlinutils.set.multiset.testutils.TestMultiSet
 import xyz.lbres.kotlinutils.set.multiset.testutils.TestMutableMultiSet
 import kotlin.test.assertEquals
