+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/const/constmutable/runMutableUnaryTests.kt
@@ -1,6 +1,6 @@
-package xyz.lbres.kotlinutils.set.multiset.constmutable
+package xyz.lbres.kotlinutils.set.multiset.const.constmutable
 
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.const.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
